Add files via upload
This is everything I did for this project.Then again,I would appreciate if you could tell me if you see that the two snakes are accelerating  and if they don't how can I fix the problem? I appreciate all help. The demands are in the word documents.I apologize it's all in french,but if you have questions about what the program is suppose to do, I can answer you!
Thank you all so much and have a good day !

-Celia